use bluebird's `fromCallback` instead of `promisify` for node-mapnik
This is a shorter and clearer method of doing it.

`fromCallback` returns a promise, it takes a function that is passed
the resolving callback.

E.g.

    var aPromise = fromCallback((cb) => map.render(opts, cb));

fromCallback, calls the arrow function with an error-first-callback
that map.render will invoke, when it does it will resolve the promis
returned by fromCallback.