Conditional If-Modified-Since caching based on datemodified from ES
This replaces the scheduled jobs with one that queries ElasticSearch
more frequently to check if anything has changed. It uses the max
datestamps per doc type to trigger recordset/indexTerms updates and
then uses the global max datemodified to set a system wide
`Last-Modified` header.

I think we could refine this further; e.g. if you are searching
publishers we could set the last modified of that result to the max
datemodified of only the publishers which is probably older than that
of records, enabling higher hit rate.