Don't package mock data into docker container
This is a couple hundred megabytes we don't need to distribute with
the container.

If we want to run the tests in an environment without the live ES
available inside of docker then it would probably be better to put
this data in a standalone data volume container and mount that into
the testing container.