Don't serve a response with cache headers before done loading data
This fixes a direct bug with how attribution is attached to a record
as well as a subtle bug where it was possible to serve a page while
recordsets are still loading, but then any future request used
If-Modified-Since and so never saw the updated info.

This blocks serving responses until caches are warmed up.