Include app start as a lastmodified date, only enable conditional-get on `/v2/...` urls
`/v2/...` urls  are the ones backed by ElasticSearch for which we can
make a guarantee that if nothing has changed in elasticsearch then the
cache copy is still valid.

The big remaining hole there is that an update to the app itself is a point
where everything could change. Add the app's start time as part of the
dates that we choose the maximum from for the Last-Modified date. It
would probably be better to use a build date, but we don't yet have a
good way of inserting that into the process.