Remove async/await from all query generating functions
There was a bug with these being inconsistently used as async that
made queries be skipped (and instead we just got everything in the
resultset).

Originally this started to be async because `checkTerms` was
async. Now we do more work to block the entire app until `checkTerms` is
ready and that function is synchronous; it had caused this type of
problem before. Recognizing that checkTerms had been the only
source (infection?) of async in this thread of code we were able to
remove that decorator off of a lot of functions.