Wait for the app on startup instead of every request
We want to make sure the app is ready (loaded indexTerms etc) before
handling any requests. Previously we did this by blocking the handling
of a given request until it was ready.

Instead we can block server startup. This way the proxy server/service
discovery manager won't try and send traffic to this server until it
is actually ready.