Fix caching to correctly apply to `/v2/` prefix
When I tried to make the conditional request only apply to the /v2/
prefix I had mistakenly thought that koa-router `.use(path, middleware)`
applied the middleware if the path was a prefix of the request.

This is wrong (and I'm not the only one:
https://github.com/alexmingoia/koa-router/issues/257); so make a new
middlware that is just a prefix matcher and applies the given
middlware, otherwise skip to the next one. Use this to apply the
lastModified conditional request guard to only the `/v2/` prefix