Don't redirect on map creation
The idea was that once created the URL with shortcode was now the
canonical place for that information so redirect there. Apparently
this violates some browser CORS issues though. Something about it not
being a simple request and it can't preflight following a redirect.

Instead just treat it as an internal redirect by calling the other
function directly.