More functional style
this changes a bunch of

   var a = [];
   list.forEach(f(e) { a.push(transform(e));})

to use map calls; I also changed  places where an object return
value is declared at the top and then amended through the function to
instead declare a couple variables in the function and return the
complete object.

I think this is a better style, uses about 10% fewer lines of code,
and chrome's tracing optimizer does better if all object properties
are defined at once.