Use sol-redis-pool for redis connection pool caching
This changes the way we expose redisclient. Instead of it being a
function that gets you a connection it now appears more like an
already estabilished connection. When you call one of the methods on
it though it acquires a conn from the pool, runs the method and then
ensures the connection is released.

I'm creating the promisifying wrappers more manually here instead of
using bluebird's promisify because I don't want to modify the original
redis library object, and I want to run extra code inline
here (acquire/release).