Allow objects in redis cache keys
if we want to cache on the query definition for example we need to
hash that into the key. However, the main key in redis should just be
a join of the strings passed in so that we get keys like
`queryShim-42a8b999fac-v1.0.0-4949944` instead of just a raw hash.