Change default color palette to "Dark2" and buckets to 5
https://www.idigbio.org/redmine/issues/2225

The previous color palette we were using 'Paired' picked as its first
color a light blue that blended too much with open street maps water
color.

This is using the "Dark2" palette from ColorBrewer which isn't
advertised as being as color-blind friend as paired is, but will
perhaps work better on our map.

We're hoping that by decreasing the buckets to 5 it will still be easy
enough to distinguish.