Rewrite map coloring code with points
One of the problems we'd been running into that I hadn't really
grokked yet was that chroma was interpolating between the colors we
wanted so that's part of why  it looked similar but everything was
muddier and off.

This just takes the ColorBrewer list and walks down the list
explicitly until it runs out of colors and uses the last color for
everything (and the 'other' bucket).

I've tried to update the tests to reflect this new reality and remove
code that tested long deprecated features (being able to set colors per
record feature)