Treat esclient as a jest mock
This changes the level at which we mock out elasticsearch
responses. Previously it was calls to searchshim that got memoized to
the FS, now we do that at the ElasticSearch client level. This allows
the searchShim code to be tested in the CI environment too.

In addition we use the files anytime they are present and write them
if they are missing and we can query ES. This makes running repeated
tests larger, but we may need to clear that cache every now and again
to test on new indexes.

This also reduces the amount of environment testing going on in the
core code; instead relying on `jest.mock` calls in the test
files. This is slightly more onerous but less complex in the
production code.