Have travis build, test, and push docker image
Previously travis tested the code in one environment and docker
hub built an image using a different base env. Now we have travis try
and build the docker image, run tests on that and then if those tests
pass upload the images to dockerhub.

My goal on image tag names is:

  * any push to a branch will create a docker image tag named after
    the branch; this is like `latest` except per branch.
  * don't tag docker images if it is a pull request
  * If the build is for a git tag then use that tag name for the
    docker image.