docs: mocks and updates/additions [skip ci]
Apparently mocks are cached Elasticsearch responses that might need to
be recreated to test for public API changes. I didn't know that.
And the rest of the team maybe might have possibly sort of also
forgot how to regenerate mocks.

So I wrote how to do so in README.md and,
for corresponding errors in doing so, in TROUBLESHOOTING.md.

Also:
- Updated node version requirement
- Documented environment variable 'IDB_REDIS_AUTH'
- Add link to Elasticsearch setup instructions